fix(pr): rename file
